+++ src/m2dir/envelope/cli.rs
@@ -15,12 +15,30 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod arg;
-pub mod cli;
-pub mod client;
-pub mod create;
-pub mod delete;
-pub mod envelope;
-pub mod flag;
-pub mod list;
-pub mod message;
+use anyhow::Result;
+use clap::Subcommand;
+use pimalaya_cli::printer::Printer;
+
+use crate::m2dir::{
+    client::M2dirClient,
+    envelope::{get::M2dirEnvelopeGetCommand, list::M2dirEnvelopeListCommand},
+};
+
+/// Manage M2DIR envelopes.
+///
+/// An envelope contains header information about a message such as
+/// date, subject, from, to, cc, bcc, etc.
+#[derive(Debug, Subcommand)]
+pub enum M2dirEnvelopeCommand {
+    Get(M2dirEnvelopeGetCommand),
+    List(M2dirEnvelopeListCommand),
+}
+
+impl M2dirEnvelopeCommand {
+    pub fn execute(self, printer: &mut impl Printer, client: M2dirClient) -> Result<()> {
+        match self {
+            Self::Get(cmd) => cmd.execute(printer, client),
+            Self::List(cmd) => cmd.execute(printer, client),
+        }
+    }
+}