+++ src/m2dir/envelope/mod.rs
@@ -15,12 +15,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod arg;
 pub mod cli;
-pub mod client;
-pub mod create;
-pub mod delete;
-pub mod envelope;
-pub mod flag;
+pub mod get;
 pub mod list;
-pub mod message;