+++ src/m2dir/message/mod.rs
@@ -15,12 +15,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod arg;
 pub mod cli;
-pub mod client;
-pub mod create;
-pub mod delete;
-pub mod envelope;
-pub mod flag;
-pub mod list;
-pub mod message;
+pub mod export;
+pub mod get;
+pub mod read;
+pub mod save;